Add volatile flag and count to examine messages
We have a looper thread filling a message queue with data from the
UART. As we process those messages we need to wait for the queue to
empty before we process the data. When the message is a simple 12
byte "response", this is easy. When there is a data packet, for
example when sending "device info" at 42 bytes, we start to see
multiple calls to process data.

To get around this, we need to wait for the buffer to stop growing,
which may be a terrible idea, but time will tell.

In order to wait responsibly we add a boolean to check if we are
already waiting (scheduled timer 500ms). If so, we don't need to
start up another timer.

Signed-off-by: Stephano Cetola <stephano.cetola@linux.intel.com>