Create a uart manager and move functionality
There's a lot of business going on with the uart. We should probably
not do that all in the main activity as it really has nothing to do
with the main activity.

This refactors out all that logic into a manager, as well as creating
the concept of a "Threaded Manager". Not sure if this is a good idea
or not yet, but we shall see.

Signed-off-by: Stephano Cetola <stephano.cetola@linux.intel.com>