Add the ability to read and write to realtime db
Add the Firebase Realtime Database to the FirestoreManager. Also, add the logic
for communicating with the ADA app. This includes a "current user" object that
can be read and written to the rtdb.

All possible values are now stored in static variables inside object that
categorize each of the sets of strings.

Signed-off-by: Stephano Cetola <stephano.cetola@linux.intel.com>